requirements.txt
Flask==3.1.2
numpy==2.2.6
opencv-python==4.13.0.90
torch==2.10.0
pillow==12.1.0
matplotlib==3.10.8